refactor: extract item key creation logic into a util
