+++ packages/elements/src/list/helpers/renderer.ts
@@ -1,7 +1,7 @@
 import type { CollectionComposer } from '@refinitiv-ui/utils/collection.js';
-import { uuid } from '@refinitiv-ui/utils/uuid.js';
 import type { Item, ItemType, ItemData } from '../../item';
 import '../../item/index.js';
+import { createKey, getItemKey } from './key.js';
 import { Renderer } from '../renderer.js';
 
 /**
@@ -16,11 +16,10 @@ type Context = {
  * This is the default renderer for lists.
  */
 export class ListRenderer extends Renderer {
-
   /**
    * Renderer key prefix, used in combination with item value to give unique id to each item
    */
-  public key: string = uuid().split('-')[0];
+  public key: string = createKey();
 
   constructor (context?: unknown) {
     /**
@@ -39,7 +38,7 @@ export class ListRenderer extends Renderer {
       el.label = composer.getItemPropertyValue(item, 'label') as string;
       el.subLabel = composer.getItemPropertyValue(item, 'subLabel') as string;
       el.value = composer.getItemPropertyValue(item, 'value') as string;
-      el.id = `${this.key}-${el.value}`;
+      item.value && (el.id = getItemKey(this.key, el.value));
       el.icon = composer.getItemPropertyValue(item, 'icon') as string;
       el.highlighted = composer.getItemPropertyValue(item, 'highlighted') === true;
       el.selected = composer.getItemPropertyValue(item, 'selected') === true;
